requirements.txt
python-dotenv
instagrapi
fastapi
uvicorn[standard]
pydantic-settings
requests
openai
